fix: update typescript to ^5.9.3, fix Buffer type strictness for TS 5.9
- Updated typescript devDep from 5.3.3 to ^5.9.3 (required by @tobilu/qmd peer dep)
- Regenerated package-lock.json for CI compatibility
- Fixed 11 Buffer.concat() and writeFile/read type errors across 7 files
- Added src/types/web-streams.d.ts for ReadableStream type stubs
- Fixed node:module createRequire → readFileSync in crash-report.ts
- Fixed srv implicit any in web.ts
- Fixed noUncheckedIndexedAccess errors in models-cmd/models

+++ src/cli/slash/paste.ts
@@ -95,7 +95,7 @@ function readClipboardImage(): ClipboardResult {
       Buffer.isBuffer(xc.stdout) &&
       xc.stdout.length > 100
     ) {
-      writeFileSync(tmp, xc.stdout);
+      writeFileSync(tmp, xc.stdout as Uint8Array);
       return {
         ok: true,
         path: tmp,
